RemoteOscilloscop/pyproject.toml
[project]
name = "remoteoscilloscope"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "jijinbei", email = "hikoharu14142@outlook.jp" }
]
dependencies = [
    "pyvisa>=1.14.1",
    "pyvisa-py>=0.7.2",
    "psutil>=6.0.0",
    "zeroconf>=0.132.2",
    "python-dotenv>=1.0.1",
    "pillow>=10.3.0",
    "pandas>=2.2.2",
    "matplotlib>=3.9.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/remoteoscilloscope"]
